requirements.txt
apify-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
requests
beautifulsoup4
dnspython
openai
python-dotenv
playwright
playwright-stealth
markdownify
